+++ src/main/kotlin/step4/domain/cell/CellType.kt
@@ -1,4 +1,4 @@
-package step4.domain
+package step4.domain.cell
 
 import java.lang.IllegalArgumentException
