+++ src/test/kotlin/step4/domain/cell/CellTest.kt
@@ -1,4 +1,4 @@
-package step4.domain
+package step4.domain.cell
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
@@ -6,10 +6,10 @@ import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.throwable.shouldHaveMessage
-import step4.domain.CellFixture.aCell
-import step4.domain.CellType.MINE
-import step4.domain.CellType.UNKNOWN
-import step4.domain.CellType.ZERO
+import step4.domain.cell.CellFixture.aCell
+import step4.domain.cell.CellType.MINE
+import step4.domain.cell.CellType.UNKNOWN
+import step4.domain.cell.CellType.ZERO
 
 class CellTest : FunSpec({
