+++ src/test/kotlin/step4/domain/state/ReadyTest.kt
@@ -5,9 +5,9 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.throwable.shouldHaveMessage
 import io.kotest.matchers.types.shouldBeTypeOf
-import step4.domain.Cell
-import step4.domain.Cells
-import step4.domain.Coordinate
+import step4.domain.cell.Cell
+import step4.domain.cell.Cells
+import step4.domain.coordinate.Coordinate
 
 class ReadyTest : FunSpec({
